pyproject.toml
[tool.poetry]
name = "capezone"
version = "0.1.0"
description = "Backend Recruiting Task"
authors = ["nandan-jain <annejain2@gmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
